Removed/moved all dead code
