+++ aws/network_acl_entry.go
@@ -58,34 +58,6 @@ func expandNetworkAclEntries(configured []interface{}, entryType string) ([]*ec2
 	return entries, nil
 }
 
-func flattenNetworkAclEntries(list []*ec2.NetworkAclEntry) []map[string]interface{} {
-	entries := make([]map[string]interface{}, 0, len(list))
-
-	for _, entry := range list {
-
-		newEntry := map[string]interface{}{
-			"from_port": *entry.PortRange.From,
-			"to_port":   *entry.PortRange.To,
-			"action":    *entry.RuleAction,
-			"rule_no":   *entry.RuleNumber,
-			"protocol":  *entry.Protocol,
-		}
-
-		if entry.CidrBlock != nil {
-			newEntry["cidr_block"] = *entry.CidrBlock
-		}
-
-		if entry.Ipv6CidrBlock != nil {
-			newEntry["ipv6_cidr_block"] = *entry.Ipv6CidrBlock
-		}
-
-		entries = append(entries, newEntry)
-	}
-
-	return entries
-
-}
-
 func protocolStrings(protocolIntegers map[string]int) map[int]string {
 	protocolStrings := make(map[int]string, len(protocolIntegers))
 	for k, v := range protocolIntegers {
