+++ aws/tagsACMPCA.go
@@ -1,9 +1,6 @@
 package aws
 
 import (
-	"log"
-	"regexp"
-
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/acmpca"
 )
@@ -50,18 +47,4 @@ func tagsToMapACMPCA(ts []*acmpca.Tag) map[string]string {
 	}
 
 	return result
-}
-
-// compare a tag against a list of strings and checks if it should
-// be ignored or not
-func tagIgnoredACMPCA(t *acmpca.Tag) bool {
-	filter := []string{"^aws:"}
-	for _, v := range filter {
-		log.Printf("[DEBUG] Matching %v with %v\n", v, aws.StringValue(t.Key))
-		if r, _ := regexp.MatchString(v, aws.StringValue(t.Key)); r == true {
-			log.Printf("[DEBUG] Found AWS specific tag %s (val: %s), ignoring.\n", aws.StringValue(t.Key), aws.StringValue(t.Value))
-			return true
-		}
-	}
-	return false
 }
