+++ aws/validators_test.go
@@ -3057,3 +3057,25 @@ func TestValidateSecretManagerSecretNamePrefix(t *testing.T) {
 		}
 	}
 }
+
+func validateEcrRepositoryName(v interface{}, k string) (ws []string, errors []error) {
+	value := v.(string)
+	if len(value) < 2 {
+		errors = append(errors, fmt.Errorf(
+			"%q must be at least 2 characters long: %q", k, value))
+	}
+	if len(value) > 256 {
+		errors = append(errors, fmt.Errorf(
+			"%q cannot be longer than 256 characters: %q", k, value))
+	}
+
+	// http://docs.aws.amazon.com/AmazonECR/latest/APIReference/API_CreateRepository.html
+	pattern := `^(?:[a-z0-9]+(?:[._-][a-z0-9]+)*/)*[a-z0-9]+(?:[._-][a-z0-9]+)*$`
+	if !regexp.MustCompile(pattern).MatchString(value) {
+		errors = append(errors, fmt.Errorf(
+			"%q doesn't comply with restrictions (%q): %q",
+			k, pattern, value))
+	}
+
+	return
+}
